100% code coverage and interface tweaks
Migrated AndSpecification and OrSpecification to the new
BinaryCompositeSpecification base class to keep things simple.  Kept the
previous CompositeSpecification as NaryCompositeSpecification.  Renamed
PredicatedSpecification to PredicateSpecification (just sounds better).
Changed the input to PredicateSpecification to be an actual Predicate
instead of a lambda function (you can pass in a lambda function, though,
so it works both ways).  Updated AssemblyInfo.cs with appropriate
information.  Version is 0.1.0.  Added and updated unit tests to reach
100% code coverage.